redesign continued

+++ src/core/structs/tree/vectors/tree_vec.rs
@@ -1,7 +1,7 @@
 use std::ops::{Index, IndexMut};
 use crate::core::structs::tree::tree_node::TreeNode;
 
-pub trait TreeVec<T>: Index<u64, Output = TreeNode<T>> + IndexMut<u64, Output = TreeNode<T>> {
+pub trait TreeVec<T>: Index<usize, Output = TreeNode<T>> + IndexMut<usize, Output = TreeNode<T>> {
     fn push(&mut self, value: T) -> i32;
     //fn pop(&mut self) -> Option<TreeNode<T>>;
     fn get(&mut self, index: i32) -> Option<TreeNode<T>>;
